everything is ready for docker

+++ grpc_store/routes/deleteProductById.js
@@ -1,13 +1,28 @@
 const {pool} = require('./../dbConfig.js')
 const grpc = require('grpc')
+const verifySession = require('./../middlewares/verify.js')
 
 module.exports = async function(call, callback) {
+
 	try {
-		const {id} = call.request
-		const {rows} = await pool.query(`DELETE FROM products WHERE id = $1`,[id])
-		callback(null, {success: true, error:{code:null, data: null}})
+		const result = await verifySession(call)
+
+		if(result.success){
+
+			const {id} = call.request
+			if(id) {
+				const {rows} = await pool.query(`DELETE FROM products WHERE id = $1`,[id])
+				callback(null, {success: true, error:{code:null, data: null}})
+			} else {
+				callback(null, {success:false, error:{code:grpc.status.INVALID_ARGUMENT, data:'argument is missing'}})
+			}
+
+		} else {
+			callback(new Error(result.data));
+		}
+
 	} catch(err) {
-		callback(null, {success:false, error:{code:grpc.status.INTERNAL, data:'internal server error'}})
+		callback(new Error("Internal server error"));
 		console.error(err)
 	}
 }
